requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
pymysql==1.1.0
pydantic==2.3.0
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0
alembic==1.12.0 